src/sage/rings/function_field/order.py: Split into modules order_...
